refactor: cleanup names
