+++ main.py
@@ -1,5 +1,5 @@
 from devportal.core import ui, plugin_loader  # noqa: F401
-from devportal.core.app import app  # noqa: F401
+from devportal.core import api  # noqa: F401
 
 # Loads all devportal plugins
 plugin_loader.run()
