+++ src/devportal/core/ui/app/read.py
@@ -1,18 +1,18 @@
-from devportal.apps import main as apps
-from devportal.core.app import app, templates
-from devportal.core.model import ApplicationBase
-from devportal.vms import main as vms
-from devportal.dbs import main as dbs
 from fastapi import Request
 from fastapi.responses import HTMLResponse, RedirectResponse
 
+from devportal.apps import main as apps
+from devportal.core import api, templates
+from devportal.vms import main as vms
+from devportal.dbs import main as dbs
 
-@app.get("/ui/app/{app_code}", response_class=RedirectResponse, status_code=301)
+
+@api.get("/ui/app/{app_code}", response_class=RedirectResponse, status_code=301)
 async def get_application_root(app_code: str):
     return f"/ui/app/{app_code}/about"
 
 
-@app.get("/ui/app/{app_code}/{partial}", response_class=HTMLResponse)
+@api.get("/ui/app/{app_code}/{partial}", response_class=HTMLResponse)
 async def get_application_partial(
     request: Request, app_code: str, partial: str = "about"
 ):
@@ -24,12 +24,12 @@ async def get_application_partial(
     )
 
 
-@app.get("/ui/content/app/{app_code}", response_class=RedirectResponse, status_code=301)
+@api.get("/ui/content/app/{app_code}", response_class=RedirectResponse, status_code=301)
 async def get_application_content_root(app_code: str):
     return f"/ui/content/app/{app_code}/about"
 
 
-@app.get("/ui/content/app/{app_code}/{partial}", response_class=HTMLResponse)
+@api.get("/ui/content/app/{app_code}/{partial}", response_class=HTMLResponse)
 async def get_application_content_partial(
     request: Request, app_code: str, partial: str
 ):
@@ -41,7 +41,7 @@ async def get_application_content_partial(
     )
 
 
-@app.get("/ui/partial/app/{app_code}/about", response_class=HTMLResponse)
+@api.get("/ui/partial/app/{app_code}/about", response_class=HTMLResponse)
 async def get_application_partial_about(request: Request, app_code: str):
     app = apps.get_app(app_code)
     return templates.TemplateResponse(
@@ -49,7 +49,7 @@ async def get_application_partial_about(request: Request, app_code: str):
     )
 
 
-@app.get("/ui/partial/app/{app_code}/servers", response_class=HTMLResponse)
+@api.get("/ui/partial/app/{app_code}/servers", response_class=HTMLResponse)
 async def get_application_partial_servers(request: Request, app_code: str):
     servers = vms.list_vms(app_code)
     return templates.TemplateResponse(
@@ -59,21 +59,11 @@ async def get_application_partial_servers(request: Request, app_code: str):
     )
 
 
-@app.get("/ui/partial/app/{app_code}/databases", response_class=HTMLResponse)
+@api.get("/ui/partial/app/{app_code}/databases", response_class=HTMLResponse)
 async def get_application_partial_databases(request: Request, app_code: str):
     db_list = dbs.list_dbs(app_code)
     return templates.TemplateResponse(
         request=request,
         name="partial/app/databases.html",
         context={"app": {"code": app_code}, "databases": db_list},
-    )
-
-
-@app.post("/ui/app", response_class=HTMLResponse, status_code=201)
-async def create_application(request: Request, app: ApplicationBase):
-    apps.create_app(app)
-    return templates.TemplateResponse(
-        request=request,
-        name="partial/app/alert.html",
-        context={"app": app, "message": "apps.alert.new.success", "level": "success"},
     )
